updated timer expires
