+++ drivers/sensor/example_sensor/example_sensor.c
@@ -15,33 +15,8 @@
 #include <zephyr/logging/log.h>
 LOG_MODULE_REGISTER(example_sensor, CONFIG_SENSOR_LOG_LEVEL);
 
-// timer expires
 // Declare the timer
 struct k_timer my_timer;
-
-// Timer expiry function
-void my_timer_expiry_function(struct k_timer *timer_id) {
-    LOG_INF("Timer expired");
-}
-
-// Timer stop function - el expires timer doesn't have stop function...
-void my_timer_stop_function(struct k_timer *timer_id) {
-    LOG_INF("Timer stopped");
-}
-
-// Initialize the timer
-void init_my_timer(void) {
-    k_timer_init(&my_timer, my_timer_expiry_function, my_timer_stop_function);
-    LOG_INF("Timer Initialized");
-}
-
-// In your function where you start the timer
-// start a periodic timer that expires once every CONFIG_MY_TIMER_PERIOD
-void start_my_timer(void) {
-    k_timer_start(&my_timer, K_SECONDS(CONFIG_MY_TIMER_INITIAL_DELAY), K_SECONDS(CONFIG_MY_TIMER_PERIOD));
-}
-
-K_TIMER_DEFINE(my_timer, my_timer_expiry_function, NULL);
 
 //static FILE *csv_file = NULL;
 struct example_sensor_data {
@@ -53,50 +28,23 @@ struct example_sensor_config {
 
 };
 
-static int example_sensor_sample_fetch(const struct device *dev, enum sensor_channel chan)
-{
+static int example_sensor_sample_fetch(const struct device *dev, enum sensor_channel chan){
+
     struct example_sensor_data *data = dev->data;
 
-    /*
-    Toma la data del sensor...
-    */
-
-    if(!data->csv_file){
+    if (!data->csv_file) {
         data->csv_file = fopen(CONFIG_PATH_TO_CSV_FILE, "r");
-        if(!data->csv_file){
+        if (!data->csv_file) {
             LOG_ERR("Error opening file");
             return -ENOENT;
         }
     }
 
-    char line[CONFIG_SIZE_LINE_MAX];
-
-    if (fgets(line, sizeof(line), data->csv_file))
-    {
-        int value;
-        if (sscanf(line, "%d", &value) == 1)
-        {
-            data->state = value;
-        }
-        else
-        {
-            LOG_ERR("Error reading file");
-            return -EIO;
-        }
-    }
-    else
-    {
-        LOG_INF("Reading file last number");
-        fclose(data->csv_file);
-        data->csv_file = NULL;
-    }
-
     return 0;
+
 }
 
-static int example_sensor_channel_get(const struct device * dev,
-                                      enum sensor_channel   chan,
-                                      struct sensor_value * val){
+static int example_sensor_channel_get(const struct device * dev, enum sensor_channel chan, struct sensor_value * val){
     struct example_sensor_data *data = dev->data;
 
     if (chan != SENSOR_CHAN_PROX && chan !=  SENSOR_CHAN_AMBIENT_TEMP) {
@@ -104,9 +52,6 @@ static int example_sensor_channel_get(const struct device * dev,
     }
 
     val->val1 = data->state;
-
-// ESTO ES PARA PUENTEAR EL SENSOR A UN VALOR 100 FIJO
-//    val->val1 = 100;
 
     return 0;
 }
@@ -122,6 +67,91 @@ static int example_sensor_init(const struct device *dev)
     data->csv_file = NULL;
     return 0;
 }
+
+// Timer expiry function
+void my_timer_expiry_function(struct k_timer *timer_id) {
+    const struct device *sensor;
+	int ret;
+	struct sensor_value value_x;
+    struct example_sensor_data *data;
+    char line[CONFIG_SIZE_LINE_MAX];
+
+    LOG_INF("Timer expired");
+
+    sensor = DEVICE_DT_GET(DT_NODELABEL(example_sensor));
+
+    if (!device_is_ready(sensor)) {
+        LOG_ERR("Sensor not ready");
+        return;
+    }
+
+    data = sensor->data;
+
+    if (fgets(line, sizeof(line), data->csv_file)) {
+        int value;
+        if (sscanf(line, "%d", &value) == 1) {
+            data->state = value;
+        } else {
+            LOG_ERR("Error reading file");
+            return;
+        }
+    } else {
+        LOG_INF("Reached end of file, resetting to start");
+        fseek(data->csv_file, 0, SEEK_SET);
+        if (fgets(line, sizeof(line), data->csv_file)) {
+            int value;
+            if (sscanf(line, "%d", &value) == 1) {
+                data->state = value;
+            } else {
+                LOG_ERR("Error reading file");
+                return;
+            }
+        } else {
+            LOG_ERR("Error resetting file");
+            return;
+        }
+    }
+
+    ret = sensor_channel_get(sensor, SENSOR_CHAN_PROX, &value_x);
+    printk("sensor_sample_get ret: %d  valor del sensor: %d\n", ret, value_x.val1);
+}
+
+// Timer stop function - el expires timer doesn't have stop function...
+void my_timer_stop_function(struct k_timer *timer_id) {
+    LOG_INF("Timer stopped");
+}
+
+// Initialize the timer
+void init_my_timer(void) {
+
+    const struct device *sensor;
+	int ret;
+	struct sensor_value value_x;
+
+    sensor = DEVICE_DT_GET(DT_NODELABEL(example_sensor));
+    if (!device_is_ready(sensor)) {
+    LOG_ERR("Sensor not ready");
+    return 0;
+    }
+
+    k_timer_init(&my_timer, my_timer_expiry_function, my_timer_stop_function);
+    LOG_INF("Timer Initialized");
+
+    ret = sensor_sample_fetch(sensor);
+    if (ret) {
+		printk("sensor_sample_fetch failed ret %d\n", ret);
+		return;
+		}
+}
+
+// In your function where you start the timer
+// start a periodic timer that expires once every CONFIG_MY_TIMER_PERIOD
+void start_my_timer(void) {
+    k_timer_start(&my_timer, K_SECONDS(CONFIG_MY_TIMER_INITIAL_DELAY), K_SECONDS(CONFIG_MY_TIMER_PERIOD));
+}
+
+K_TIMER_DEFINE(my_timer, my_timer_expiry_function, NULL);
+
 
 #define EXAMPLE_SENSOR_INIT(i)						       					\
     static struct example_sensor_data example_sensor_data_##i;	       		\
